+++ packages/sphinx-test-reports/tests/test_custom_template.py
@@ -1,10 +1,17 @@
-from sphinx_testing import with_app
+from pathlib import Path
+
+import pytest
 
 
-@with_app(buildername="html", srcdir="doc_test/custom_tr_template")
-def test_custom_template(app, status, warning):
+@pytest.mark.parametrize(
+    "test_app",
+    [{"buildername": "html", "srcdir": "doc_test/custom_tr_template"}],
+    indirect=True,
+)
+def test_custom_template(test_app):
+    app = test_app
     app.build()
-    html = (app.outdir / "index.html").read_text()
+    html = Path(app.outdir / "index.html").read_text()
 
     # assert changed order when creating a testreport html page
     pos1 = html.index("<strong>Imported data</strong>")